refactor config structure and add mypy

+++ apps/mobsinet/simulator/models/nodes/packet.py
@@ -6,19 +6,17 @@ if TYPE_CHECKING:
     from .abc_node import AbcNode
 
 
-
 class Packet:
-    
+
     def __init__(self, message: 'AbcMessage'):
         self.message = message
-        self.origin: 'AbcNode' 
-        self.destination: 'AbcNode' 
+        self.origin: 'AbcNode'
+        self.destination: 'AbcNode'
         self.type: Literal["UNICAST", "MULTICAST", "DUMMY"]
-        self.positive_delivery: bool 
-        self.arriving_time: int = 0
-        self.sending_time: int = 0
+        self.positive_delivery: bool
+        self.arriving_time: float = 0
+        self.sending_time: float = 0
         self.intensity: float = 0
-        
 
     def set_message(self, message: 'AbcMessage'):
         self.message = message
@@ -27,8 +25,7 @@ class Packet:
         """Deny the delivery of the packet."""
 
         self.positive_delivery = False
-        
-    
+
     def to_json(self):
         return {
             'message': self.message.__str__(),
